Create README.md
I am a Rusty Engineer from 2014, living in Ludhiana aiming for 2026 Phd. I am interested in Quantum AI and believes it has potential to create wonders. Derived by these thoughts I want to improve my rusty programming skills and writing codes with the help of AI for now while I am teaching myself Python. I used to hate coding but gone are the days when it was not needed as much now its a necessity.